Clean up the system tests
