+++ DevToolkt/coreSystemTests/src/commonTest/kotlin/dev/toolkt/core/platform/PlatformWeakMapTests.kt
@@ -1,10 +1,12 @@
 package dev.toolkt.core.platform
 
+import dev.toolkt.core.platform.test_utils.assertEqualsEventually
+import dev.toolkt.core.platform.test_utils.runTestDefault
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
+import kotlinx.coroutines.test.TestResult
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.withContext
-import kotlin.math.roundToInt
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -12,22 +14,22 @@ import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 
-private class Key(
-    @Suppress("unused") val data: List<Int>,
-) {
-    companion object {
-        private val random = Random(seed = 0)
+private class GarbageCollectionTest {
+    // A key of random data that weights roughly 32 KiB
+    private class Key(
+        @Suppress("unused") val data: List<Int>,
+    ) {
+        companion object {
+            private val random = Random(seed = 0)
 
-        fun create(): Key = Key(
-            data = List(8000) { random.nextInt() },
-        )
+            fun create(): Key = Key(
+                data = List(8000) { random.nextInt() },
+            )
+        }
+
+        override fun toString(): String = "Key#${hashCode()}"
     }
 
-    override fun toString(): String = "Key#${hashCode()}"
-}
-
-
-private class GarbageCollectionTest {
     private val key1: Key = Key.create()
 
     private var key2: Key? = Key.create()
@@ -37,8 +39,6 @@ private class GarbageCollectionTest {
     private var key4: Key? = Key.create()
 
     suspend fun run() {
-        println("Starting the PlatformWeakMap system test...")
-
         val weakMap = mutableWeakMapOf<Key, String>()
 
         weakMap[key1] = "value1"
@@ -70,56 +70,9 @@ private class GarbageCollectionTest {
 
 class PlatformWeakMapSystemTests {
     @Test
-    fun testGarbageCollection() = runTest(
-        timeout = 15.seconds,
+    fun testGarbageCollection() = runTestDefault(
+        duration = 15.seconds,
     ) {
-        withContext(Dispatchers.Default) {
-            GarbageCollectionTest().run()
-        }
+        GarbageCollectionTest().run()
     }
-}
-
-private enum class WaitUntilResult {
-    Timeout, Success,
-}
-
-private suspend fun <T> assertEqualsEventually(
-    pauseDuration: Duration,
-    timeoutDuration: Duration,
-    expected: T,
-    actual: () -> T,
-) {
-    val result = waitUntil(
-        pauseDuration = pauseDuration,
-        timeoutDuration = timeoutDuration,
-    ) {
-        actual() == expected
-    }
-
-    when (result) {
-        WaitUntilResult.Timeout -> {
-            throw AssertionError("Expected value to eventually become $expected, but it is at ${actual()} after checking every $pauseDuration for $timeoutDuration")
-        }
-
-        WaitUntilResult.Success -> {}
-    }
-}
-
-private suspend fun waitUntil(
-    pauseDuration: Duration,
-    timeoutDuration: Duration,
-    predicate: () -> Boolean,
-): WaitUntilResult {
-    val tryCount = (timeoutDuration / pauseDuration).roundToInt()
-
-    (tryCount downTo 0).forEach { tryIndex ->
-        delay(pauseDuration)
-        PlatformSystem.collectGarbage()
-
-        if (predicate()) {
-            return WaitUntilResult.Success
-        }
-    }
-
-    return WaitUntilResult.Timeout
 }
