+++ DevToolkt/coreSystemTests/src/commonTest/kotlin/dev/toolkt/core/platform/test_utils/waitUntilTestUtils.kt
@@ -1,0 +1,29 @@
+package dev.toolkt.core.platform.test_utils
+
+import dev.toolkt.core.platform.PlatformSystem
+import kotlinx.coroutines.delay
+import kotlin.math.roundToInt
+import kotlin.time.Duration
+
+enum class WaitUntilResult {
+    Timeout, Success,
+}
+
+suspend fun waitUntil(
+    pauseDuration: Duration,
+    timeoutDuration: Duration,
+    predicate: () -> Boolean,
+): WaitUntilResult {
+    val tryCount = (timeoutDuration / pauseDuration).roundToInt()
+
+    (tryCount downTo 0).forEach { tryIndex ->
+        delay(pauseDuration)
+        PlatformSystem.collectGarbage()
+
+        if (predicate()) {
+            return WaitUntilResult.Success
+        }
+    }
+
+    return WaitUntilResult.Timeout
+}
